+++ packages/hydrogen/src/components/Link/Link.client.tsx
@@ -1,0 +1,1 @@
+export {Link} from 'react-router-dom';